refactor: general requirements and renaming
