+++ requirements.txt
@@ -6,6 +6,7 @@ numpy==1.20.3
 pandas==1.2.4
 Pillow==8.2.0
 pyparsing==2.4.7
+pytest==6.2.4
 python-dateutil==2.8.1
 pytz==2021.1
 scikit-learn==0.24.2
